Fix color problem when printing NUL-terminated strings
This commit fixes a problem in which changing colors didn't affect
the color of printed NUL-terminated strings.

The problem was caused because we changed ds before calling the function.
We did that so we can print heap strings (we changed ds to be equal to the segment
of the heap string).
We had no problem with that until we accessed the global color variable
from the data segment inside the print function.
That created a problem because the data segment was the heap segment
and not the regular data segment, resulting in a read of garbage.