Remove string in expr instead of object
String objects shouild not free the string source because
when creating more than 1 instance of a string from the same expr,
you double-free (or even more times) the string source when you destruct
the string objects.

Instead, the string objects should have a "copy" pointer
to the source in the expr.

This commit also makes us free the string in the expr.